Adding copyright header and removing unused code
