+++ app/cdap/components/GridTable/DirectiveComponents/SendToError/index.tsx
@@ -1,13 +1,29 @@
+/*
+ * Copyright © 2022 Cask Data, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not
+ * use this file except in compliance with the License. You may obtain a copy of
+ * the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations under
+ * the License.
+ */
+
 import { FormControl, FormGroup, Typography, MenuItem } from '@material-ui/core';
 import React, { useState, useEffect } from 'react';
-import { SET_VARIABLE_NAME, SELECT_ROW_WHERE, SEND_TO_ERROR, IGNORE_CASE } from '../constants';
+import { SEND_TO_ERROR, IGNORE_CASE } from '../constants';
 import { SEND_TO_ERROR_OPTIONS, FILTER_PLACEHOLDER } from '../options';
 import { useStyles } from '../styles';
 import FormInputFieldComponent from '../ParseComponent/FormInputFieldComponent';
 import SelectOptionComponent from '../HashComponent/SelectComponent';
 import InputCheckbox from '../../../ParsingDrawer/Components/InputCheckbox';
 
-const DefineVariable = (props) => {
+const SendToErrorComponent = (props) => {
   const { setDirectiveComponentsValue, directiveComponentValues } = props;
   const [filterCondition, setFilterCondition] = useState('EMPTY');
   const [ignoreCase, setIgnoreCase] = useState(false);
@@ -86,4 +102,4 @@ const DefineVariable = (props) => {
   );
 };
 
-export default DefineVariable;
+export default SendToErrorComponent;
