+++ app/cdap/components/GridTable/DirectiveComponents/UsingPattern_Extract/index.tsx
@@ -1,12 +1,20 @@
-import {
-  FormControl,
-  FormControlLabel,
-  FormGroup,
-  Radio,
-  RadioGroup,
-  MenuItem,
-} from '@material-ui/core';
-import InputCheckbox from '../../../ParsingDrawer/Components/InputCheckbox';
+/*
+ * Copyright © 2022 Cask Data, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not
+ * use this file except in compliance with the License. You may obtain a copy of
+ * the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations under
+ * the License.
+ */
+
+import { FormControl, FormGroup, MenuItem } from '@material-ui/core';
 import React, { useState, useEffect } from 'react';
 import {
   SELECT_PATTERN,
@@ -22,7 +30,7 @@ import FormInputFieldComponent from '../ParseComponent/FormInputFieldComponent';
 import SelectOptionComponent from '../HashComponent/SelectComponent';
 import ParseComponent from '../ParseComponent';
 
-const FilterComponent = (props) => {
+const UsingPattern = (props) => {
   const { setDirectiveComponentsValue, directiveComponentValues } = props;
   const [patternName, setPatternName] = useState('');
   const [startValue, setStartValue] = useState('');
@@ -142,4 +150,4 @@ const FilterComponent = (props) => {
   );
 };
 
-export default FilterComponent;
+export default UsingPattern;
